notes/2-docker-aws/docker-exercise/requirements.txt
Flask==1.0.2
flask_restplus==0.12.1
keras==2.2.4
tensorflow==1.14.0
Pillow==5.1.0
werkzeug==0.16.0
numpy==1.17.2